Mark WARNING and higher as cold
In GCC and Clang `__attribute__((cold))` has effects in two places:

1. The code generated for the cold function is optimized for size and
   not speed and the code generate for it may be placed far away from
   the main line of the program in the build object. In this case it
   should have no effect as all the functions are trivially elidable.
2. Any branch which inevitably leads to calling a cold function is also
   considered cold. This is the effect this patch is aiming to activate.

Consider this code fragment:
```
if (some_c_call(...) == 0) {
  // Branch A
  LOG(INFO) << "It worked, proceeding as normal";
  // The processing I would like to be fast
  return;
}
// Branch B
PLOG(ERROR) << "Something went wrong";
// Error handling code
```

From playing around on godbolt Clang consistently presumes that Branch A
is statically less likely (possibly because it's only taken for one
result of a possible 4 billion) and emits code that is most performant
when Branch B is taken. Whereas my hope as the programmer is that
`some_c_call()` will usually return zero and I would like Branch A to be
favoured.

More seriously, the compiler spends a lot of resources optimizing the
speed of Branch B. This negatively impacts compile times and output
object size. It would be more appropriate to spend less time optimizing
the bad cases and to make the error-path code as small as possible to
minimize overall binary size.